requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
plotly==5.19.0
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
